--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "workbench.colorCustomizations": {
-        "activityBar.background": "#113515",
-        "titleBar.activeBackground": "#174A1D",
-        "titleBar.activeForeground": "#F5FCF6"
-    }
-}
