+++ package.json
@@ -1,23 +1,20 @@
 {
-    "name": "gmail_login_ensure",
+    "name": "rust-books",
     "type": "module",
     "version": "1.0.0",
-    "main": "index.js",
+    "main": "main.ts",
     "license": "MIT",
     "scripts": {
-        "start": "node dist/main",
-        "watch": "tsc --watch"
+        "start": "tsx main.ts"
     },
     "dependencies": {
-        "@types/lodash": "^4.14.176",
-        "async-lock": "^1.3.0",
-        "chalk": "^4.1.2",
-        "delay": "^5.0.0",
-        "lodash": "^4.17.21",
         "meow": "^10.1.1",
-        "playwright": "^1.18.0-alpha-jan-8-2022",
-        "source-map-support": "^0.5.20",
+        "p-limit": "^5.0.0",
+        "playwright": "1.42",
         "toml": "^3.0.0",
-        "typescript": "^4.4.4"
+        "tsx": "^4.7.2"
+    },
+    "devDependencies": {
+        "@types/node": "^20.12.7"
     }
 }
