+++ tsconfig.json
@@ -8,17 +8,10 @@
         "esModuleInterop": true,
         "inlineSourceMap": true,
         "skipLibCheck": true,
-        "lib": [
-            "esnext",
-            "dom"
-        ],
+        "lib": ["esnext", "dom"],
         "outDir": "./dist",
-        "rootDir": "./src",
+        "rootDir": "./src"
     },
-    "exclude": [
-        "node_modules"
-    ],
-    "files": [
-        "src/main.ts",
-    ]
+    "exclude": ["node_modules"],
+    "files": ["main.ts"]
 }
